Move reduce op tests to accelerator tests
