+++ .azure/hpu-tests.yml
@@ -107,16 +107,6 @@ jobs:
 
     - bash: |
         python -m pytest -sv \
-          test_pytorch/test_reduce_op.py -k strategy\
-          --hpus 8 --junitxml=hpu-reduce-op-strategy-results.xml
-        python -m pytest -sv \
-          test_pytorch/test_reduce_op.py -k logging\
-          --hpus 8 --junitxml=hpu-reduce-op-logging-results.xml
-      workingDirectory: tests/
-      displayName: 'Reduce Op tests'
-
-    - bash: |
-        python -m pytest -sv \
           test_pytorch/test_datamodule.py \
           test_pytorch/test_profiler.py \
           --forked --hpus 1 --junitxml=hpu1_test-results.xml
@@ -125,6 +115,12 @@ jobs:
 
     - bash: |
         set -ex
+        python -m pytest -sv \
+          test_pytorch/test_accelerator.py -k reduce_op_strategy \
+          --hpus 8 --junitxml=hpu8-reduce-op-strategy-results.xml
+        python -m pytest -sv \
+          test_pytorch/test_accelerator.py -k reduce_op_logging \
+          --hpus 8 --junitxml=hpu8-reduce-op-logging-results.xml
         python -m pytest -sv  test_pytorch/test_accelerator.py \
           --forked --hpus 8 --junitxml=hpu8_accel_test-results.xml
         python -m pytest -sv test_pytorch/test_profiler.py \
