+++ tests/test_pytorch/test_accelerator.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 import os
 from unittest import mock
+from contextlib import contextmanager
+from typing import Any, Optional, Union
 
 import pytest
 import torch
@@ -24,13 +26,16 @@ if module_available("lightning"):
     from lightning.pytorch import Callback, Trainer, seed_everything
     from lightning.pytorch.demos.boring_classes import BoringModel
     from lightning.pytorch.utilities.exceptions import MisconfigurationException
+    from lightning.fabric.utilities.types import ReduceOp
 elif module_available("pytorch_lightning"):
     from pytorch_lightning import Callback, Trainer, seed_everything
     from pytorch_lightning.demos.boring_classes import BoringModel
     from pytorch_lightning.utilities.exceptions import MisconfigurationException
+    from lightning_fabric.utilities.types import ReduceOp
 
 from lightning_habana.pytorch.accelerator import HPUAccelerator
 from lightning_habana.pytorch.strategies import HPUParallelStrategy, SingleHPUStrategy
+from lightning_habana.pytorch.plugins import HPUPrecisionPlugin
 from tests.helpers import ClassifDataModule, ClassificationModel
 
 # The external strategies, accelerator gets registerd with trainer initialization
@@ -53,7 +58,8 @@ def test_device_name():
 
 
 def test_accelerator_selected():
-    trainer = Trainer(accelerator=HPUAccelerator(), strategy=SingleHPUStrategy())
+    trainer = Trainer(accelerator=HPUAccelerator(),
+                      strategy=SingleHPUStrategy())
     assert isinstance(trainer.accelerator, HPUAccelerator)
 
 
@@ -179,7 +185,8 @@ def test_stages_correct(tmpdir):
 
 
 def test_accelerator_is_hpu():
-    trainer = Trainer(accelerator=HPUAccelerator(), devices=1, strategy=SingleHPUStrategy())
+    trainer = Trainer(accelerator=HPUAccelerator(),
+                      devices=1, strategy=SingleHPUStrategy())
     assert isinstance(trainer.accelerator, HPUAccelerator)
     assert trainer.num_devices == 1
 
@@ -211,7 +218,8 @@ def test_accelerator_auto_with_devices_hpu():
 
 
 def test_strategy_choice_single_strategy():
-    trainer = Trainer(strategy=SingleHPUStrategy(device=torch.device("hpu")), accelerator=HPUAccelerator(), devices=1)
+    trainer = Trainer(strategy=SingleHPUStrategy(
+        device=torch.device("hpu")), accelerator=HPUAccelerator(), devices=1)
     assert isinstance(trainer.strategy, SingleHPUStrategy)
 
     trainer = Trainer(accelerator="hpu", devices=1)
@@ -221,7 +229,8 @@ def test_strategy_choice_single_strategy():
 @pytest.mark.skipif(device_count() <= 1, reason="Test requires multiple HPU devices")
 def test_strategy_choice_parallel_strategy():
     trainer = Trainer(
-        strategy=HPUParallelStrategy(parallel_devices=[torch.device("hpu")] * 8),
+        strategy=HPUParallelStrategy(
+            parallel_devices=[torch.device("hpu")] * 8),
         accelerator=HPUAccelerator(),
         devices=8,
     )
@@ -330,3 +339,172 @@ def test_hpu_device_stats_monitor():
     ]
     for f in fields:
         assert any(f in h for h in hpu_stats)
+
+
+@contextmanager
+def does_not_raise():
+    """No-op context manager as a complement to pytest.raises."""
+    yield
+
+
+class BaseBM(BoringModel):
+    """Model to test with reduce ops."""
+
+    def __init__(self, reduce_op=None):
+        """Init."""
+        super().__init__()
+        self.reduce_op = reduce_op
+        self.logged_value_start = 42
+
+    def training_step(self, batch, batch_idx):
+        """Training step."""
+        # Each ddp process logs 3 values: 42, 43, and 44.
+        # logger performs reduce depending on the reduce_op
+        loss = super().training_step(batch, batch_idx)
+        self.log(
+            "logged_value",
+            self.logged_value_start,
+            prog_bar=True,
+            sync_dist=True,
+            reduce_fx=self.reduce_op,
+            on_epoch=True,
+        )
+        self.logged_value_start += 1
+        return loss
+
+
+class MetricsCallback(Callback):
+    """PyTorch Lightning metric callback."""
+
+    def __init__(self):
+        """Init."""
+        super().__init__()
+        self.metrics = []
+
+    def on_validation_epoch_end(self, trainer, pl_module):
+        """Copy trainer metrics."""
+        metric = copy.deepcopy(trainer.logged_metrics)
+        self.metrics.append(metric)
+
+
+class MockHPUParallelStrategy(HPUParallelStrategy):
+    def __init__(
+        self,
+        reduce_op="sum",
+        **kwargs: Any,
+    ):
+        super().__init__()
+        self.reduce_op = reduce_op
+        self.logged_messages = []
+
+    def reduce(
+        self, tensor: torch.Tensor, group: Optional[Any] = None, reduce_op: Optional[Union[ReduceOp, str]] = "sum"
+    ) -> torch.Tensor:
+        return super().reduce(tensor, group, self.reduce_op)
+
+
+@pytest.mark.skipif(HPUAccelerator.auto_device_count() <= 1, reason="Test requires multiple HPU devices")
+def test_hpu_parallel_reduce_op_strategy_default():
+    """Test default reduce_op."""
+    strategy = MockHPUParallelStrategy()
+    # Assert that the strategy's reduce_op attribute is set to the default "sum"
+    assert strategy.reduce_op == "sum"
+
+
+@pytest.mark.skipif(HPUAccelerator.auto_device_count() < 8, reason="Test requires multiple HPU devices")
+@pytest.mark.parametrize(
+    ("reduce_op", "expectation"),
+    [
+        ("sum", does_not_raise()),
+        ("max", does_not_raise()),
+        ("min", does_not_raise()),
+        ("mean", does_not_raise()),
+        (
+            "product",
+            pytest.raises(
+                AssertionError,
+                match="Unsupported ReduceOp product. Only 'sum', 'min', and 'max' are supported with HCCL",
+            ),
+        ),
+        (ReduceOp.SUM, does_not_raise()),
+        (ReduceOp.MIN, does_not_raise()),
+        (ReduceOp.MAX, does_not_raise()),
+        (ReduceOp.AVG, does_not_raise()),
+        (
+            ReduceOp.PRODUCT,
+            pytest.raises(
+                AssertionError,
+                match="Unsupported ReduceOp RedOpType.PRODUCT. Only 'sum', 'min', and 'max' are supported with HCCL",
+            ),
+        ),
+    ],
+    ids=[
+        "sum",
+        "max",
+        "min",
+        "mean",
+        "product",
+        "ReduceOp.SUM",
+        "ReduceOp.MIN",
+        "ReduceOp.MAX",
+        "ReduceOp.AVG",
+        "ReduceOp.PRODUCT",
+    ],
+)
+def test_reduce_op_strategy(tmpdir, reduce_op, expectation):
+    """Tests all reduce in HPUParallel strategy."""
+    seed_everything(42)
+    _model = BoringModel()
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        accelerator=HPUAccelerator(),
+        devices=8,
+        strategy=MockHPUParallelStrategy(
+            reduce_op=reduce_op, start_method="spawn"),
+        max_epochs=1,
+        fast_dev_run=3,
+    )
+    with expectation:
+        trainer.fit(_model)
+
+
+@pytest.mark.skipif(HPUAccelerator.auto_device_count() < 8, reason="Test requires multiple HPU devices")
+@pytest.mark.parametrize(
+    ("reduce_op", "logged_value_epoch", "logged_value_step"),
+    [
+        # Epoch = Sum(42, 43, 44) * 8, Step = 44 * 8 (for 8 ddp processes)
+        ("sum", 1032.0, 352.0),
+        # Epoch = Max(42, 43, 44), Step = Max(44, ... (x8))
+        ("max", 44.0, 44.0),
+        # Epoch = Min(42, 43, 44), Step = Min(44, ... (x8))
+        ("min", 42.0, 44.0),
+        # Epoch = Mean(42(x8), 43(x8), 44(x8)), Step = Mean(44, ... (x8))
+        ("mean", 43.0, 44.0),
+    ],
+)
+def test_reduce_op_logging(tmpdir, reduce_op, logged_value_epoch, logged_value_step):
+    """Test reduce_op with logger and sync_dist."""
+    # Logger has its own reduce_op sanity check.
+    # It only accepts following string reduce_ops {min, max, mean, sum}
+    # Each ddp process logs 3 values: 42, 43, and 44.
+    # logger performs reduce depending on the reduce_op
+    seed_everything(42)
+    _model = BaseBM(reduce_op=reduce_op)
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        accelerator=HPUAccelerator(),
+        devices=8,
+        strategy=HPUParallelStrategy(start_method="spawn"),
+        max_epochs=1,
+        fast_dev_run=3,
+        plugins=HPUPrecisionPlugin(precision="bf16-mixed"),
+        callbacks=[MetricsCallback()],
+    )
+    trainer.fit(_model)
+
+    assert torch.allclose(
+        trainer.callback_metrics.get("logged_value_epoch"), torch.tensor(logged_value_epoch), atol=1e-4
+    )
+    assert torch.allclose(trainer.callback_metrics.get(
+        "logged_value_step"), torch.tensor(logged_value_step), atol=1e-4)
